[Update] and refactor

+++ src/Console/InstallCommand.php
@@ -3,6 +3,7 @@
 namespace Sajadsdi\LaravelDynamicRouter\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
 
 class InstallCommand extends Command
@@ -38,25 +39,30 @@ class InstallCommand extends Command
     private function installRoutes()
     {
         $this->comment('Adding web-routes in web route ...');
+        
+        
+        if (File::exists(base_path('routes/web.php'))) {
+            //require web-route.php in web.php
+            $webRout = file_get_contents(base_path('routes/web.php'));
 
-        //require web-route.php in web.php
-        $webRout = file_get_contents(base_path('routes/web.php'));
-
-        if (Str::contains($webRout, "require('web-routes.php');")) {
-            $this->warn("require('web-routes.php'); is exists in web route ............ SKIPPED");
-        }else {
-            file_put_contents(base_path('routes/web.php'), $webRout . "\n\nrequire('web-routes.php');");
-            $this->info("require('web-routes.php'); added to web route ..... DONE");
+            if (Str::contains($webRout, "require('web-routes.php');")) {
+                $this->warn("require('web-routes.php'); is exists in web route ............ SKIPPED");
+            } else {
+                file_put_contents(base_path('routes/web.php'), $webRout . "\n\nrequire('web-routes.php');");
+                $this->info("require('web-routes.php'); added to web route ..... DONE");
+            }
         }
 
-        //require api-route.php in api.php
-        $apiRout = file_get_contents(base_path('routes/api.php'));
+        if (File::exists(base_path('routes/api.php'))) {
+            //require api-route.php in api.php
+            $apiRout = file_get_contents(base_path('routes/api.php'));
 
-        if (Str::contains($apiRout, "require('api-routes.php');")) {
-            $this->warn("require('api-routes.php'); is exists in web route ............ SKIPPED");
-        }else {
-            file_put_contents(base_path('routes/api.php'), $apiRout . "\n\nrequire('api-routes.php');");
-            $this->info("require('api-routes.php'); added to web route ..... DONE");
+            if (Str::contains($apiRout, "require('api-routes.php');")) {
+                $this->warn("require('api-routes.php'); is exists in web route ............ SKIPPED");
+            } else {
+                file_put_contents(base_path('routes/api.php'), $apiRout . "\n\nrequire('api-routes.php');");
+                $this->info("require('api-routes.php'); added to web route ..... DONE");
+            }
         }
     }
 }
